chore: update unit tests and utils
