+++ package.json
@@ -2,6 +2,7 @@
   "name": "new-hanafuda",
   "private": true,
   "version": "1.0.0",
+  "type": "module",
   "scripts": {
     "test": "vitest run",
     "test:watch": "vitest",
@@ -21,6 +22,7 @@
     "@heroicons/vue": "^2.0.18",
     "@nuxt/devtools": "latest",
     "@nuxt/image": "^1.0.0-rc.1",
+    "@nuxt/test-utils": "^3.19.2",
     "@nuxtjs/i18n": "^9.5.6",
     "@nuxtjs/partytown": "^1.3.0",
     "@nuxtjs/tailwindcss": "^6.8.0",
@@ -29,20 +31,21 @@
     "@playwright/test": "^1.40.1",
     "@tailwindcss/container-queries": "^0.1.1",
     "@tailwindcss/forms": "^0.5.4",
-    "@types/node": "^18.17.0",
-    "@vue/test-utils": "^2.4.1",
+    "@types/node": "^20.19.4",
+    "@vitejs/plugin-vue": "^6.0.0",
+    "@vue/test-utils": "^2.4.6",
     "@vueuse/core": "^10.2.1",
     "@vueuse/nuxt": "^10.2.1",
     "autoprefixer": "^10.4.14",
-    "happy-dom": "^9.20.3",
-    "nuxt": "^3.6.5",
+    "happy-dom": "^18.0.1",
+    "nuxt": "^3.13.0",
     "nuxt-headlessui": "^1.1.4",
     "nuxt-icon": "^0.4.2",
-    "nuxt-vitest": "^0.10.1",
+    "playwright-core": "^1.53.2",
     "postcss": "^8.4.27",
     "ts-node": "latest",
     "typescript": "latest",
-    "vitest": "^0.33.0"
+    "vitest": "^3.2.4"
   },
   "dependencies": {
     "@firebase/app-types": "^0.9.0",
